feat(shiftModel): display current month in shiftModel route
