+++ lib/utils.js
@@ -72,6 +72,27 @@ export function getCalUrl ({
   }
 }
 
+/**
+ * Get the url to display today.
+ * @param {object}   param args
+ * @param {string}   param.shiftModel   Current shift model.
+ * @param {number[]} param.today        Today array.
+ * @param {number}   param.group        Group to display.
+ */
+export function getTodayUrl ({ shiftModel, today, group }) {
+  const groupArg = group && group > 0
+    ? `?group=${group}`
+    : ''
+
+  if (today) {
+    const month = String(today[1]).padStart(2, '0')
+    const day = String(today[2]).padStart(2, '0')
+    return `/cal/${shiftModel}${groupArg}#${today[0]}-${month}-${day}`
+  } else {
+    return `/cal/${shiftModel}${groupArg}`
+  }
+}
+
 const supportZones = Info.features().zones
 /**
  * Get today as an array of [yyyy, m, d, h] (all numbers).
