+++ pages/cal/[shiftModel].jsx
@@ -5,24 +5,89 @@ This Source Code Form is subject to the terms of the Mozilla Public License, v. 
 the MPL was not distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 */
 
-import { useEffect } from 'react'
 import { useRouter } from 'next/router'
+import { DateTime, Info } from 'luxon'
 
-import { getToday, getCalUrl } from '../../lib/utils'
+import ByMonths from '../../components/by-month'
+import Downloader from '../../components/download'
+import Footer from '../../components/footer'
+import Head from '../../components/head'
+import { useTodayZeroIndex } from '../../hooks/time'
+import { shiftModelText } from '../../lib/constants'
+import { parseNumber } from '../../lib/utils'
 
+/**
+ * Route that always displays today.
+ */
 export default function ShiftModel () {
   const router = useRouter()
+  const today = useTodayZeroIndex()
+  const shiftModel = router.query.shiftModel
+  const group = parseNumber(router.query.group, 0)
+  const year = today[0]
+  const month = today[1]
 
-  useEffect(() => {
-    const today = getToday()
-    router.replace(getCalUrl({
-      isFullYear: false,
-      group: router.query.group || 0,
-      shiftModel: router.query.shiftModel,
-      year: today[0],
-      month: today[1],
-    }))
-  })
-  
-  return null
+  return (
+    <main className='flex flex-col content-center'>
+      <Head
+        title={`Monat ${year}-${
+          String(month + 1).padStart(2, '0')
+        } - ${
+          shiftModelText[shiftModel]
+        }`}
+      />
+
+      <ByMonths
+        shiftModel={shiftModel}
+        group={group}
+        search={null}
+        year={year}
+        month={month}
+        today={today}
+      />
+
+      <Downloader shiftModel={shiftModel} />
+
+      <Footer />
+    </main>
+  )
+}
+
+/**
+ * Get the props for server side rendering.
+ * @param {import('next').GetServerSidePropsContext} context Next SSR context.
+ * @returns {import('next').GetServerSidePropsResult}
+ */
+export async function getServerSideProps (context) {
+  let maxAge = 60
+
+  if (Info.features().zones) {
+    // get the diff in seconds to the next shift start
+    const now = DateTime.local().setZone('Europe/Berlin')
+
+    let hour = 6 // get next shift start
+    if (now.hour >= 22) {
+      hour = 6
+    } else if (now.hour >= 14) {
+      hour = 22
+    } else if (now.hour >= 6) {
+      hour = 14
+    }
+
+    let nextShift = now.set({
+      hour,
+      minute: 0,
+      second: 0,
+      millisecond: 0
+    })
+    if (nextShift.diff(now, 'minutes').toObject().minutes < 0) {
+      nextShift = nextShift.plus({ days: 1 })
+    }
+    maxAge = nextShift.diff(now, 'seconds').toObject().seconds
+  }
+
+  context.res.setHeader('Cache-Control', 's-maxage=' + maxAge)
+  return {
+    props: context.query
+  }
 }
